pyproject.toml
[project]
name = "yazot-mcp-server"
version = "0.1.0"
description = "Yet Another Zotero MCP server"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=3.1.0,<4.0.0",
    "pyzotero>=1.11.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "pypdf>=4.0.0",
    "beautifulsoup4>=4.14.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["yazot"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM", "RUF"]
ignore = ["E501"]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101", "F841"]

[tool.mypy]
python_version = "3.12"
plugins = ["pydantic.mypy"]
strict = true
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
disallow_untyped_defs = true
disallow_any_unimported = false
no_implicit_optional = true
check_untyped_defs = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
follow_imports = "normal"

[[tool.mypy.overrides]]
module = "pydantic"
ignore_missing_imports = false

[[tool.mypy.overrides]]
module = "pydantic.*"
ignore_missing_imports = false



[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
timeout = 60
markers = [
    "unit: pure logic, mocks only, safe to parallelize",
    "e2e: Zotero web API via MCP Client",
    "external: real external API calls (Crossref, etc.)",
    "live: requires running local Zotero instance",
]

[tool.black]
line-length = 100
target-version = ['py312']

[dependency-groups]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.6.0",
    "mypy>=1.18.2",
    "black>=24.8.0",
    "pre-commit>=3.8.0",
    "ipython>=8.0.0",
    "pytest-timeout>=2.4.0",
    "pytest-xdist>=3.5.0",
]
